LEGACY/MAQUETADO/requirements.txt
python-docx
openai
pytest
python-dotenv
PyYAML
spacy